server/requirements.txt
langchain>=0.3.0
langchain-openai>=0.0.2
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 